src/process_history/requirements.txt
boto3>=1.20.0
pyspark>=3.1.1
awsglue-local>=0.1.0
pandas>=1.3.0
